// modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/types/Space.ts
/**
 * SPDX-FileCopyrightText: (c) 2025 Liferay, Inc. https://liferay.com
 * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
 */

export type Space = {
	creatorUserId: string;
	externalReferenceCode: string;
	id: string;
	name: string;
	settings?: {logoColor: string};
};